backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.24.0
sqlalchemy[asyncio]==2.0.23
asyncpg==0.29.0
python-socketio==5.10.0
httpx==0.25.2
apscheduler==3.10.4
python-dotenv==1.0.0
pydantic-settings==2.1.0
google-generativeai==0.3.1
pydantic==2.5.3
aiohttp